Move dataset tests to a separate file (#74)
